refactor(*): move config file to `config` directory
Files located in `example` dir may be removed by file pruners (such as `yarn autoclean`,
`node-prune`, `mod-clean`). This can lead to fatal error in application.
