+++ instructions.js
@@ -14,7 +14,7 @@ const path = require('path')
 module.exports = async function (cli) {
   try {
     await cli.copy(
-      path.join(__dirname, './example/cors.js'),
+      path.join(__dirname, './config/cors.js'),
       path.join(cli.helpers.configPath(), 'cors.js')
     )
     cli.command.completed('create', 'config/cors.js')
